software-engineering/python/py-spy/_src/sandbox/pyproject.toml
[project]
name = "torch-sandbox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytorch-triton>=3.3.0",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "py-spy>=0.3.14",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
pytorch-triton = { index = "pytorch-cu124" }
torch = { index = "pytorch-cu124" }
torchvision = { index = "pytorch-cu124" }
